requirements.txt
# Core dependencies
Jinja2>=3.1.2
pyyaml>=6.0
pytest>=7.3.1
python-dotenv>=1.0.0  # For loading environment variables

# Email handling
exchangelib>=5.0.0  # For Exchange integration, if needed
pywin32>=305  # For Outlook integration

# Optional but recommended
black>=23.3.0  # Code formatting
flake8>=6.0.0  # Linting
isort>=5.12.0  # Import sorting
pre-commit>=3.3.2  # Pre-commit hooks
mypy>=1.3.0  # Type checking
